refactor(generator): split into separate files each handling part of the generation and parsing
This should provide more code readability and easier understanding of tool functionality
